requirements.txt
beautifulsoup4
boto3
elasticsearch-dsl>=6.0.0,<7.0.0
flask
flask-cors
lxml
numpy
pandas
pika
pillow
git+https://github.com/aboSamoor/polyglot.git@master
psycopg2-binary
python-levenshtein
pycountry
pymarc
pyyaml
redis
requests
requests_oauthlib
sklearn
sqlalchemy
waitress